requirements.txt
Flask
openai
numpy
youtube_transcript_api
scikit-learn
tiktoken
python-dotenv
llama_index
flask_cors
fitz